Fix getPartition overload ambiguity with clang-cl
Use an explicitly unsigned zero so MSVC-compatible overload resolution selects the partition-index overload without changing behavior on other compilers.

+++ lib/Dialect/TritonGPU/Transforms/WarpSpecialization/PartitionLoops.cpp
@@ -396,7 +396,7 @@ LogicalResult triton::gpu::partitionLoop(scf::ForOp loop) {
     return success();
 
   auto numPartitions = partitions.getNumPartitions();
-  auto defaultPartition = partitions.getPartition((int)0);
+  auto defaultPartition = partitions.getPartition(0u);
   auto loopVarCategories = classifyLoopVars(loop, defaultPartition, partitions);
   auto [loopVarIndices, newResultIndices] =
       getLoopVarIndicesToKeep(loop, defaultPartition, loopVarCategories);
